Refactored the User controller Redirected to the user profile after the file upload.

+++ src/main/java/com/ansr/service/UserService.java
@@ -2,6 +2,8 @@ package com.ansr.service;
 
 import java.util.List;
 
+import com.ansr.exceptions.common.ApplicationException;
+import com.ansr.exceptions.UserErrorCodes;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -23,10 +25,32 @@ public class UserService {
 	}
 
 	public User getUser(String userId) {
-		return userRepository.findOne(userId);
+
+		User user = userRepository.findOne(userId);
+		if(user == null) {
+			throw new ApplicationException(UserErrorCodes.USER_NOT_FOUND);
+		}
+
+		return user;
 	}
 
 	public void delete(String userId) {
+
+		User user = userRepository.findOne(userId);
+		if(user == null) {
+			throw new ApplicationException(UserErrorCodes.USER_NOT_FOUND);
+		}
+
 		userRepository.delete(userId);
+	}
+
+	public void updateUser(String userId, User user) {
+
+		User existingUser = userRepository.findOne(userId);
+		if(existingUser == null) {
+			throw new ApplicationException(UserErrorCodes.USER_NOT_FOUND);
+		}
+
+		userRepository.save(user);
 	}
 }
